it works :)

+++ internal/manager.go
@@ -132,7 +132,7 @@ type ManagerConfiguration struct {
 	VirtualShards struct {
 		Enabled bool  `json:"enabled" yaml:"enabled"`
 		Count   int32 `json:"count" yaml:"count"`       // Number of virtual shards to use
-		DmShard int   `json:"dm_shard" yaml:"dm_shard"` // Shard to use for DMs and non identifiable events
+		DmShard int32 `json:"dm_shard" yaml:"dm_shard"` // Shard to use for DMs and non identifiable events
 	} `json:"virtual_shards" yaml:"virtual_shards"`
 
 	Rest struct {
@@ -332,8 +332,8 @@ func (mg *Manager) ConsumerShardCount() int32 {
 }
 
 // GetShardIdOfGuild returns the shard id of a guild
-func (mg *Manager) GetShardIdOfGuild(guildID discord.Snowflake, shardCount int32) int64 {
-	return (int64(guildID) >> 22) % int64(shardCount)
+func (mg *Manager) GetShardIdOfGuild(guildID discord.Snowflake, shardCount int32) int32 {
+	return int32((int64(guildID) >> 22) % int64(shardCount))
 }
 
 // RoutePayloadToConsumer routes a SandwichPayload to its corresponding consumer modifying the payload itself
@@ -356,10 +356,10 @@ func (mg *Manager) RoutePayloadToConsumer(payload *sandwich_structs.SandwichPayl
 		payload.Metadata.Shard = [3]int32{}
 	} else if payload.EventDispatchIdentifier.GuildID != nil && *payload.EventDispatchIdentifier.GuildID != 0 {
 		virtualShardId := mg.GetShardIdOfGuild(*payload.EventDispatchIdentifier.GuildID, mg.Configuration.VirtualShards.Count)
-		payload.Metadata.Shard = [3]int32{0, int32(virtualShardId), int32(mg.Configuration.VirtualShards.Count)}
+		payload.Metadata.Shard = [3]int32{0, virtualShardId, mg.Configuration.VirtualShards.Count}
 	} else {
 		// Not globally routed + no guild id means it's a DM
-		payload.Metadata.Shard = [3]int32{0, int32(mg.Configuration.VirtualShards.DmShard), int32(mg.Configuration.VirtualShards.Count)}
+		payload.Metadata.Shard = [3]int32{0, mg.Configuration.VirtualShards.DmShard, mg.Configuration.VirtualShards.Count}
 	}
 
 	return nil
